% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{aims_data_doi_get}
\alias{aims_data_doi_get}
\title{AIMS Dataset DOI retriever}
\usage{
aims_data_doi_get(target)
}
\arguments{
\item{target}{A \code{\link[base]{character}} vector of length 1
specifying the dataset. Only \code{weather} or
\code{temp_loggers} are currently allowed.
#'}
}
\value{
A \code{\link[base]{character}} vector
containing the dataset DOI string.
}
\description{
Returns DOI for a given dataset
}
\examples{
\dontrun{
library(dataaimsr)
weather_doi  <-  aims_data_doi_get('weather')
ssts_doi  <-  aims_data_doi_get('temp_loggers')
}
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
