% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimsdataplatform.R
\name{aims_data_get}
\alias{aims_data_get}
\title{Get some data and return a data frame}
\usage{
aims_data_get(doi, filters = NULL, ...)
}
\arguments{
\item{doi}{Description of parameter}

\item{filters}{Description of parameter}

\item{...}{Additional arguments to be passed to internal functions
\code{\link{page_data_get}} and \code{\link{next_page_data_get}}}
}
\value{
A \code{\link[jsonlite]{fromJSON}} simplified \code{data.frame}.
}
\description{
General description here
}
\details{
Details about this function.
}
\examples{
\dontrun{
library(dataaimsr)
# assumes that user already has API key saved to
# .Renviron
weather  <-  getOption("dataaimsr.weather")
results <- aims_data_get(weather,
                    api_key = NULL,
                    filters = list('series' = 64,
                                   'from-date' = '2018-01-01',
                                   'thru-date' = '2018-01-10'))
}

}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
