% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimsdataplatform.R
\name{api_key_find}
\alias{api_key_find}
\title{Title here}
\usage{
api_key_find(api_key)
}
\arguments{
\item{api_key}{An API key obtained from
\href{https://aims.github.io/data-platform/key-request}{AIMS DataPlatform}.}
}
\value{
Either a \code{character} vector API key found
in .Renviron or, if missing entirely, a \code{warning} message.
}
\description{
General description here
}
\details{
Details about this function.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
