% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{data_doi}
\alias{data_doi}
\title{AIMS Dataset DOI retriever}
\usage{
data_doi(target)
}
\arguments{
\item{target}{A \code{\link[base]{character}} vector of length 1 specifying
the dataset. Only \code{weather} or \code{temp_loggers} are currently
allowed.}
}
\value{
A \code{\link[base]{character}} vector
containing the dataset DOI string.
}
\description{
Returns DOI for a given dataset
}
\examples{
\dontrun{
library(dataaimsr)
weather_doi <- data_doi("weather")
ssts_doi <- data_doi("temp_loggers")
}
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
\keyword{internal}
