% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr-package.R
\docType{package}
\name{dataaimsr-package}
\alias{dataaimsr-package}
\alias{dataaimsr}
\title{The 'dataaimsr' package.}
\description{
dataaimsr is the Australian Institute of Marine Science (AIMS)
Data Platform R package, and provides the user with easy access to datasets
from the AIMS Data Platform API. Please see ?aims_data for more details.
}
\references{
Australian Institute of Marine Science (AIMS). (2017). AIMS Sea Water
Temperature Observing System (AIMS Temperature Logger Program)
https://doi.org/10.25845/5b4eb0f9bb848

Australian Institute of Marine Science (AIMS). (2017). Northern Australia
Automated Marine Weather and Oceanographic Stations,
https://doi.org/10.25845/5c09bf93f315d
}
