% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimsdf-methods.R
\name{extract_map_coord}
\alias{extract_map_coord}
\title{extract_map_coord}
\usage{
extract_map_coord(x, ...)
}
\arguments{
\item{x}{An sfc_POINT}

\item{...}{Additional argument "pos" to internal function}
}
\description{
Internal
}
