% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimsdataplatform.R
\name{filter_values_get}
\alias{filter_values_get}
\title{Retrieve existing list of data attributes}
\usage{
filter_values_get(doi, filter_name)
}
\arguments{
\item{doi}{Description of parameter}

\item{filter_name}{Description of parameter}
}
\value{
A \code{\link[jsonlite]{fromJSON}} simplified \code{data.frame}.
}
\description{
General description here
}
\details{
Details about this function.
}
\examples{
\dontrun{
library(dataaimsr)
weather  <-  getOption("dataaimsr.weather")
filter_values_get(weather, filter_name = "series")
}

}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
