% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_api_key}
\alias{find_api_key}
\title{AIMS API Key retriever}
\usage{
find_api_key(api_key)
}
\arguments{
\item{api_key}{An AIMS Data Platform
\href{https://open-aims.github.io/data-platform/key-request}{API Key}}
}
\value{
Either a \code{\link[base]{character}} vector API Key found
in .Renviron or, if missing entirely, an error message.
}
\description{
This function tries to search for an API Key
}
\details{
The AIMS Data Platform R Client provides easy access to
data sets for R applications to the
\href{https://open-aims.github.io/data-platform/}{AIMS Data Platform API}.
The AIMS Data Platform requires an API Key for requests, which can
be obtained at this
\href{https://open-aims.github.io/data-platform/key-request}{link}.
It is preferred that API Keys are not stored in code. We recommend
storing the environment variable \code{AIMS_DATAPLATFORM_API_KEY}
permanently under the user's \code{.Renviron} file in order to load
the API Key automatically.

There are two types of data currently available through the
\href{https://open-aims.github.io/data-platform/}{AIMS Data Platform API}:
\href{https://weather.aims.gov.au/#/overview}{Weather} and
\href{https://tinyurl.com/h93mcojk}{Sea Water Temperature Loggers}.
They are searched internally via unique DOI identifiers.
Only one data type at a time can be passed to the argument \code{target}.

A list of arguments for \code{filters} can be exposed for both
\href{https://weather.aims.gov.au/#/overview}{Weather} and
\href{https://weather.aims.gov.au/#/overview}{Sea Water Temperature Loggers}
using function \code{\link{aims_expose_attributes}}.

Note that at present the user can inspect the range of dates for
the temperature loggers data only (see usage of argument \code{summary} in
the examples below). For that, the argument \code{summary} must be either
the string \code{"summary-by-series"} or \code{"summary-by-deployment"}.
In those cases, time filters will be ignored.

Details about available dates for each dataset and time series can be
accessed via Metadata on
\href{https://open-aims.github.io/data-platform/}{AIMS Data Platform API}.
We raise this caveat here because these time boundaries are very important;
data are collected at very small time intervals, a window of just a few days
can yield very large datasets. The query will return and error
if it reaches the system's memory capacity.

For that same reason, from version 1.1.0 onwards, we are offering the
possibility of downloading a mean daily aggregated version. For that, the
user must set \code{summary = "daily"}. In this particular case, query filter
will be taken into account.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
\keyword{internal}
