% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{format_update}
\alias{format_update}
\title{Format \code{\link[jsonlite]{fromJSON}} output list}
\usage{
format_update(results, doi)
}
\arguments{
\item{results}{A \code{\link[jsonlite]{fromJSON}} list
generated by \code{\link{json_results}}.}

\item{doi}{A \href{https://www.doi.org/}{Digital Object Identifier}
for a chosen \href{https://aims.github.io/data-platform}{AIMS data series}}
}
\value{
A \code{\link[jsonlite]{fromJSON}} \code{\link[base]{list}}
containing four elements:
\code{metadata}: a \href{https://www.doi.org/}{DOI} link
containing the metadata record for the data series;
\code{citation}: the citation information for the particular dataset;
\code{links}: the link from which the data query was retrieved;
\code{data}: an output \code{\link[base]{data.frame}}.
}
\description{
When \code{\link[jsonlite]{fromJSON}} returns a list,
format list names
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
