% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{handle_error}
\alias{handle_error}
\title{\code{\link[httr]{GET}} error handler}
\usage{
handle_error(dt_req)
}
\arguments{
\item{dt_req}{An URL \code{\link[httr]{GET}} output}
}
\value{
A \code{\link[base]{character}} vector conveying the error message.
}
\description{
Displays error status
}
\details{
This function retrieves the
status and content of \code{dt_req} via the
\pkg{httr} package.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
