% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimsdf-class.R
\name{is_aimsdf}
\alias{is_aimsdf}
\title{Checks if argument is a \code{aimsdf} object}
\usage{
is_aimsdf(x)
}
\arguments{
\item{x}{An \R object}
}
\description{
Checks if argument is a \code{aimsdf} object
}
