% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_base_end_pt}
\alias{make_base_end_pt}
\title{Expose available query filters}
\usage{
make_base_end_pt(doi, aims_version = NA)
}
\arguments{
\item{doi}{A \href{https://www.doi.org/}{Digital Object Identifier}
for a chosen
\href{https://open-aims.github.io/data-platform}{AIMS data series}}

\item{aims_version}{A \code{\link[base]{character}} string
defining the version of database. Must be "/v1.0" or "-v2.0".
If none is provided, then "-v2.0" (the most recent) is used.}
}
\description{
Expose available query filters which are allowed to be parsed either
via argument \code{summary} or \code{filters} in \code{\link{aims_data}}
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
\keyword{internal}
