% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimsdf-methods.R
\name{print.aimsdf}
\alias{print.aimsdf}
\title{print.aimsdf}
\usage{
\method{print}{aimsdf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{aimsdf}} as
returned by \code{\link{aims_data}}.}

\item{...}{Not used.}
}
\value{
A list containing a summary of the model fit as returned a
brmsfit for each model.
}
\description{
print.aimsdf
}
