% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{process_request}
\alias{process_request}
\title{Format \code{\link{json_results}} output}
\usage{
process_request(dt_req, next_page = FALSE, ...)
}
\arguments{
\item{dt_req}{An URL \code{\link[httr]{GET}} output}

\item{next_page}{Logical. Is this a multi-url request?}

\item{...}{Additional arguments to be passed to internal function
\code{\link{update_format}}}
}
\value{
A \code{\link[jsonlite]{fromJSON}} \code{\link[base]{list}}
containing four elements:
\itemize{
\item{\code{metadata}}{a \href{https://www.doi.org/}{DOI} link
containing the metadata record for the data series.}
\item{\code{citation}}{the citation information for the particular
dataset.}
\item{\code{links}}{the link from which the data query was retrieved.}
\item{\code{data}}{an output \code{\link[base]{data.frame}}.}
}
}
\description{
Wrapper function
}
\details{
This function checks for errors in \code{dt_req}
data request and processes result via
\code{\link{json_results}}.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
