% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{request_warning}
\alias{request_warning}
\title{Title here}
\usage{
request_warning(dt_req)
}
\arguments{
\item{dt_req}{Description of parameter}
}
\value{
A warning message which retrieves the \pkg{httr}
  status and content from \code{dt_req}.
}
\description{
General description here
}
\details{
Details about this function.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
