% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimsdf-methods.R
\name{summary.aimsdf}
\alias{summary.aimsdf}
\title{summary.aimsdf}
\usage{
\method{summary}{aimsdf}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{aimsdf}} as
returned by \code{\link{aims_data}}.}

\item{...}{Unused.}
}
\value{
A list containing summary info from the input data.frame.
}
\description{
summary.aimsdf
}
