% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimsdf-methods.R
\name{tail.aimsdf}
\alias{tail.aimsdf}
\title{tail.aimsdf}
\usage{
\method{tail}{aimsdf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{aimsdf}} as
returned by \code{\link{aims_data}}.}

\item{...}{Not used.}
}
\description{
tail.aimsdf
}
