% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_format}
\alias{update_format}
\title{Format \code{\link[jsonlite]{fromJSON}} output list}
\usage{
update_format(results, doi)
}
\arguments{
\item{results}{A \code{\link[jsonlite]{fromJSON}} list
generated by \code{\link{json_results}}.}

\item{doi}{A \href{https://www.doi.org/}{Digital Object Identifier}
for a chosen
\href{https://open-aims.github.io/data-platform}{AIMS data series}}
}
\value{
If \code{summary} is passed as an additional argument (see code
example below), \code{aims_data} returns a \code{\link[base]{data.frame}}
containing the summary information for the dataset (NB: currently,
\code{summary} only works for the temperature logger dataset).

Otherwise, if actual data is downloaded, a \code{\link[jsonlite]{fromJSON}}
\code{\link[base]{list}} containing three elements:
\itemize{
\item{\code{metadata}}{a \href{https://www.doi.org/}{DOI} link
containing the metadata record for the data series.}
\item{\code{citation}}{the citation information for the particular
dataset.}
\item{\code{data}}{an output \code{\link[base]{data.frame}}.}
}
}
\description{
When \code{\link[jsonlite]{fromJSON}} returns a list, format list names
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
