% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{aims_data}
\alias{aims_data}
\title{Request data via the AIMS Data Platform API}
\usage{
aims_data(doi, filters = NULL, ...)
}
\arguments{
\item{doi}{A \href{https://www.doi.org/}{Digital Object Identifier}
for a chosen \href{https://aims.github.io/data-platform}{AIMS data series}}

\item{filters}{A \code{\link[base]{list}} containing a set of
filters for the data query (see Details)}

\item{...}{Additional arguments to be passed to internal functions
\code{\link{page_data}} and \code{\link{next_page_data}}}
}
\value{
A \code{\link[jsonlite]{fromJSON}} \code{\link[base]{list}}
containing four elements:
\code{metadata}: a \href{https://www.doi.org/}{DOI} link
containing the metadata record for the data series;
\code{citation}: the citation information for the particular dataset;
\code{links}: the link from which the data query was retrieved;
\code{data}: an output \code{\link[base]{data.frame}}.
}
\description{
A function that communicates with the
the \href{https://aims.github.io/data-platform}{AIMS Data Platform}
via the AIMS Data Platform API
}
\details{
The AIMS Data Platform R Client provides easy access to
data sets for R applications to the
\href{https://aims.github.io/data-platform}{AIMS Data Platform API}.
The AIMS Data Platform requires an API Key for requests, which can
be obtained at this \href{https://aims.github.io/data-platform/key-request}{link}.
It is preferred that API Keys are not stored in code. We recommend
storing the environment variable \code{AIMS_DATAPLATFORM_API_KEY}
permanently under the user's \code{.Renviron} file in order to load
the API key automatically.

There are two types of data currently available through the
\href{https://aims.github.io/data-platform}{AIMS Data Platform API}:
\href{https://aims.github.io/data-platform/weather/index}{Weather} and
\href{https://aims.github.io/data-platform/temperature-loggers/index}{Sea Water Temperature Loggers}.
They are searched internally via unique DOI identifiers which
can be obtained by the function \code{\link{aims_data_doi}} (see Examples).
Only one DOI at a time can be passed to the argument \code{doi}.

A list of arguments for \code{filters} can be found for both
\href{https://aims.github.io/data-platform/weather/index}{Weather} and
\href{https://aims.github.io/data-platform/weather/index}{Sea Water Temperature Loggers}.

Note that currently \code{from-date} and
\code{thru-date} cannot be inspected directly, for details about dates
of available different time series can be accessed via Metadata on
\href{https://aims.github.io/data-platform}{AIMS Data Platform API}. Despite
this limitation, these time boundaries are very important because the
data are collected at very small time intervals, so just a few days of
time interval can yield massive datasets. The query will return and error
if it reaches the system's memory capacity.
}
\examples{
\dontrun{
library(dataaimsr)
# assumes that user already has API key saved to
# .Renviron
weather_doi  <-  aims_data_doi("weather")
ssts_doi     <-  aims_data_doi("temp_loggers")

# 64 corresponds to air temperature from Davies Reef
wdata <- aims_data(weather_doi,
                       api_key = NULL,
                       filters = list("series" = 64,
                                      "from-date" = "2018-01-01",
                                      "thru-date" = "2018-01-10"))

# Downloads Chlorophyll data from all sites between given time interval
cdata <- aims_data(weather_doi,
                       api_key = NULL,
                       filters = list("parameters" = "Chlorophyll",
                                      "from-date" = "2003-01-01",
                                      "thru-date" = "2003-01-02"))

sdata <- aims_data(ssts_doi,
                       api_key = NULL,
                       filters = list("site-name" = "Bickerton Island"))
}

}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
