% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aims_data.R
\name{aims_data}
\alias{aims_data}
\title{Request data via the AIMS Data Platform API}
\usage{
aims_data(doi, filters = NULL, ...)
}
\arguments{
\item{doi}{A \href{https://www.doi.org/}{Digital Object Identifier}
for a chosen
\href{https://open-aims.github.io/data-platform}{AIMS data series}}

\item{filters}{A \code{\link[base]{list}} containing a set of
filters for the data query (see Details)}

\item{...}{Additional arguments to be passed to internal functions
\code{\link{page_data}} and \code{\link{next_page_data}}}
}
\value{
If \code{summary} is passed as an additional argument (see code
example below), \code{aims_data} returns a \code{\link[base]{data.frame}}
containing the summary information for the dataset (NB: currently,
\code{summary} only works for the temperature logger dataset).

Otherwise, if actual data is downloaded, a \code{\link[jsonlite]{fromJSON}}
\code{\link[base]{list}} containing three elements:
\itemize{
\item{\code{metadata}}{a \href{https://www.doi.org/}{DOI} link
containing the metadata record for the data series.}
\item{\code{citation}}{the citation information for the particular
dataset.}
\item{\code{data}}{an output \code{\link[base]{data.frame}}.}
}
}
\description{
A function that communicates with the
the \href{https://open-aims.github.io/data-platform}{AIMS Data Platform}
via the AIMS Data Platform API
}
\details{
The AIMS Data Platform R Client provides easy access to
data sets for R applications to the
\href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}.
The AIMS Data Platform requires an API Key for requests, which can
be obtained at this
\href{https://open-aims.github.io/data-platform/key-request}{link}.
It is preferred that API Keys are not stored in code. We recommend
storing the environment variable \code{AIMS_DATAPLATFORM_API_KEY}
permanently under the user's \code{.Renviron} file in order to load
the API Key automatically.

There are two types of data currently available through the
\href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}:
\href{https://weather.aims.gov.au/#/overview}{Weather} and
\href{https://tinyurl.com/h93mcojk}{Sea Water Temperature Loggers}.
They are searched internally via unique DOI identifiers which
can be obtained by the function \code{\link{aims_data_doi}} (see Examples).
Only one DOI at a time can be passed to the argument \code{doi}.

A list of arguments for \code{filters} can be exposed for both
\href{https://weather.aims.gov.au/#/overview}{Weather} and
\href{https://weather.aims.gov.au/#/overview}{Sea Water Temperature Loggers}
using function \code{\link{expose_attributes}}.

Note that at present the user can inspect the range of dates for
the temperature loggers data only (see usage of argument \code{summary} in
\code{\link{aims_data}}). Details about available dates for each dataset
and time series can be accessed via Metadata on
\href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}.
We raise this caveat here because these time boundaries are very important;
data are collected at very small time intervals, so just a few days of
time interval can yield massive datasets. The query will return and error
if it reaches the system's memory capacity.
}
\examples{
\dontrun{
library(dataaimsr)
# assumes that user already has API key saved to
# .Renviron

weather_doi <- aims_data_doi("weather")
ssts_doi <- aims_data_doi("temp_loggers")

# start downloads:
# 1. downloads weather data from
# site Yongala
# within a defined date range
wdf_a <- aims_data(weather_doi, api_key = NULL,
                   filters = list(site = "Yongala",
                                  from_date = "2018-01-01",
                                  thru_date = "2018-01-02"))$data

# 2. downloads weather data from all sites
# under series_id 64 from Davies Reef
# within a defined date range
wdf_b <- aims_data(weather_doi, api_key = NULL,
                   filters = list(series_id = 64,
                                  from_date = "1991-10-18",
                                  thru_date = "1991-10-19"))$data
head(wdf_b)
range(wdf_b$time)

# 3. downloads weather data from all sites
# under series_id 64 from Davies Reef
# within defined date AND time range
wdf_c <- aims_data(weather_doi, api_key = NULL,
                   filters = list(series_id = 64,
                                  from_date = "1991-10-18T06:00:00",
                                  thru_date = "1991-10-18T12:00:00"))$data
head(wdf_c)
range(wdf_c$time)

# 4. downloads all parameters from all sites
# within a defined date range
wdf_d <- aims_data(weather_doi, api_key = NULL,
                   filters = list(from_date = "2003-01-01",
                                  thru_date = "2003-01-02"))$data
# note that there are multiple sites and series
# so in this case, because we did not specify a specific
# parameter, series within sites could differ by both
# parameter and depth
head(wdf_d)
unique(wdf_d[, c("site", "series_id", "series")])
unique(wdf_d$parameter)
range(wdf_d$time)

# 5. downloads chlorophyll from all sites
# within a defined date range
wdf_e <- aims_data(weather_doi, api_key = NULL,
                   filters = list(parameter = "Chlorophyll",
                                  from_date = "2018-01-01",
                                  thru_date = "2018-01-02"))$data
# note again that there are multiple sites and series
# however in this case because we did specify a specific
# parameter, series within sites differ by depth only
head(wdf_e)
unique(wdf_e[, c("site", "series_id", "series", "depth")])
unique(wdf_e$parameter)
range(wdf_e$time)

# 6. downloads temperature data
# summarised by series
sdf_a <- aims_data(ssts_doi, api_key = NULL, summary = "summary-by-series")
head(sdf_a)
dim(sdf_a)

# 7. downloads temperature data
# summarised by series
# for all sites that contain data
# within a defined date range
sdf_b <- aims_data(ssts_doi, api_key = NULL, summary = "summary-by-series",
                   filters = list("from_date" = "2018-01-01",
                                  "thru_date" = "2018-12-31"))
head(sdf_b)
dim(sdf_b) # a subset of sdf_a

# 8. downloads temperature data
# summarised by deployment
sdf_c <- aims_data(ssts_doi, api_key = NULL,
                   summary = "summary-by-deployment")
head(sdf_c)
dim(sdf_c)
}

}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
