% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{api_key_find}
\alias{api_key_find}
\title{AIMS API Key retriever}
\usage{
api_key_find(api_key)
}
\arguments{
\item{api_key}{An API Key obtained from
\href{https://aims.github.io/data-platform/key-request}{AIMS DataPlatform}.}
}
\value{
Either a \code{\link[base]{character}} vector API Key found
in .Renviron or, if missing entirely, an error message.
}
\description{
This function tries to search for an API Key
}
\details{
The AIMS Data Platform R Client provides easy access to
data sets for R applications to the
\href{https://aims.github.io/data-platform}{AIMS Data Platform API}.
The AIMS Data Platform requires an API Key for requests, which can
be obtained at this \href{https://aims.github.io/data-platform/key-request}{link}.
It is preferred that API Keys are not stored in code. We recommend
storing the environment variable \code{AIMS_DATAPLATFORM_API_KEY}
permanently under the user's \code{.Renviron} file in order to load
the API Key automatically.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
