% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{expose_attributes}
\alias{expose_attributes}
\title{Expose available query filters}
\usage{
expose_attributes(verbose = TRUE, ...)
}
\arguments{
\item{verbose}{Should explanations be printed to screen? Defaults to TRUE.}

\item{...}{unused.}
}
\value{
A \code{\link[base]{list}} of a \code{\link[base]{character}} vector
each: one detailing summary modes, another detailing filters.
}
\description{
Expose available query filters which are allowed to be parsed either
via argument \code{summary} or \code{filters} in \code{\link{aims_data}}
}
\details{
Use this function to learn which summary modes and
filters are allowed. The summary option is only available to
the temperature logger data. Future versions will implement the
same feature for the weather station data.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
