% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_values.R
\name{filter_values}
\alias{filter_values}
\title{Retrieve vector of existing filter values}
\usage{
filter_values(target, filter_name)
}
\arguments{
\item{target}{A \code{\link[base]{character}} vector of length 1 specifying
the dataset. Only \code{weather} or \code{temp_loggers} are currently
allowed.}

\item{filter_name}{A \code{\link[base]{character}} string containing the
name of the filter. Must be "site", "subsite", "series", or
"parameter". See details.}
}
\value{
Either a \code{\link[base]{data.frame}} if
\code{filter_name = "series"}, else a \code{\link[base]{character}}
vector.
}
\description{
This is a utility function which allows to user
to query about the existing possibilities of
a given filter name
}
\details{
For a full description of each valid filter_name see
?\code{\link{expose_attributes}}. In the temperature logger dataset,
"subsite" is equivalent to "series"; moreover, note that there is only one
parameter being measured (i.e. water temperature), so the "parameter" filter
contains one single value.
}
\examples{
\dontrun{
library(dataaimsr)
filter_values("weather", filter_name = "site")
filter_values("temp_loggers", filter_name = "subsite")
}

}
\seealso{
\code{\link{aims_data}}, \code{\link{expose_attributes}}
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
