% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{filter_values_get}
\alias{filter_values_get}
\title{Retrieve existing list of data attributes}
\usage{
filter_values_get(doi, filter_name)
}
\arguments{
\item{doi}{A \href{https://www.doi.org/}{Digital Object Identifier}
for a chosen \href{https://aims.github.io/data-platform}{AIMS data series}}

\item{filter_name}{A \code{\link[base]{character}} vector containing the name of
the filter (see Details)}
}
\value{
Either a \code{\link[base]{data.frame}} or a \code{\link[base]{character}} vector.
}
\description{
This is a utility function which allows to user
to query about the existing possibilities of 
a given filter name
}
\details{
A list of arguments for \code{filters} can be found for both
[Weather](https://aims.github.io/data-platform/weather/index) and 
[Sea Water Temperature Loggers](https://aims.github.io/data-platform/weather/index).
}
\examples{
\dontrun{
library(dataaimsr)
weather_doi  <-  getOption("dataaimsr.weather")
ssts_doi     <-  getOption("dataaimsr.temp_loggers")
filter_values_get(weather_doi, filter_name = "sites")
filter_values_get(weather_doi, filter_name = "series")
filter_values_get(weather_doi, filter_name = "parameters")
filter_values_get(ssts_doi, filter_name = "sites")
}

}
\seealso{
\code{\link{aims_data_get}}
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
