% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{format_update}
\alias{format_update}
\title{Title here}
\usage{
format_update(results)
}
\arguments{
\item{results}{A \code{\link[jsonlite]{fromJSON}} simplified
\code{data.frame} generated from internal function
\code{\link{json_results}}.}
}
\value{
A \code{\link[jsonlite]{fromJSON}} simplified \code{data.frame}.
}
\description{
General description here
}
\details{
Details about this function.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
