% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{json_results}
\alias{json_results}
\title{\code{\link[jsonlite]{fromJSON}} data request}
\usage{
json_results(dt_req)
}
\arguments{
\item{dt_req}{An URL \code{\link[httr]{GET}} output}
}
\value{
\code{aims_data} returns a \code{\link[base]{data.frame}} of class
\code{\link{aimsdf}}.

If \code{summary} is passed as an additional argument,
the output shows the summary information for the target dataset (i.e.
weather or temperature loggers)
(NB: currently, \code{summary} only works for the temperature logger
database). If \code{summary} is \emph{not} passed as an additional argument, then
the output contains monitoring data.
The output also contains five attributes (empty strings if
\code{summary} is passed as an additional argument):
\itemize{
\item{\code{metadata}}{a \href{https://www.doi.org/}{DOI} link
containing the metadata record for the data series.}
\item{\code{citation}}{the citation information for the particular
dataset.}
\item{\code{parameters}}{The measured parameters comprised in the
output.}
\item{\code{type}}{The type of dataset. Either "monitoring" if
\code{summary} is not specified, or a "summary-by-" otherwise.}
\item{\code{target}}{The input target.}
}
}
\description{
Wrapper function
}
\details{
This function submits a \code{dt_req} data request via
\code{\link[jsonlite]{fromJSON}}.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
\keyword{internal}
