% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimsdataplatform.R
\name{json_results}
\alias{json_results}
\title{Title here}
\usage{
json_results(dt_req)
}
\arguments{
\item{dt_req}{Description of parameter}
}
\value{
A \code{\link[jsonlite]{fromJSON}} simplified \code{data.frame}.
}
\description{
General description here
}
\details{
Details about this function.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
