% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_data.R
\name{next_page_data}
\alias{next_page_data}
\title{Further data requests via the AIMS Data Platform API}
\usage{
next_page_data(url, api_key = NULL, ...)
}
\arguments{
\item{url}{A data retrieval URL}

\item{api_key}{An AIMS Data Platform
\href{https://open-aims.github.io/data-platform/key-request}{API Key}}

\item{...}{Additional arguments to be passed to internal function
\code{\link{update_format}}}
}
\value{
If \code{summary} is passed as an additional argument (see code
example below), \code{aims_data} returns a \code{\link[base]{data.frame}}
containing the summary information for the dataset (NB: currently,
\code{summary} only works for the temperature logger dataset).

Otherwise, if actual data is downloaded, a \code{\link[jsonlite]{fromJSON}}
\code{\link[base]{list}} containing three elements:
\itemize{
\item{\code{metadata}}{a \href{https://www.doi.org/}{DOI} link
containing the metadata record for the data series.}
\item{\code{citation}}{the citation information for the particular
dataset.}
\item{\code{data}}{an output \code{\link[base]{data.frame}}.}
}
}
\description{
Similar to \code{\link{page_data}}, but for cases #' where there are
multiple URLs for data retrieval
}
\details{
The AIMS Data Platform R Client provides easy access to
data sets for R applications to the
\href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}.
The AIMS Data Platform requires an API Key for requests, which can
be obtained at this
\href{https://open-aims.github.io/data-platform/key-request}{link}.
It is preferred that API Keys are not stored in code. We recommend
storing the environment variable \code{AIMS_DATAPLATFORM_API_KEY}
permanently under the user's \code{.Renviron} file in order to load
the API Key automatically.

There are two types of data currently available through the
\href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}:
\href{https://weather.aims.gov.au/#/overview}{Weather} and
\href{https://tinyurl.com/h93mcojk}{Sea Water Temperature Loggers}.
They are searched internally via unique DOI identifiers which
can be obtained by the function \code{\link{aims_data_doi}} (see Examples).
Only one DOI at a time can be passed to the argument \code{doi}.

A list of arguments for \code{filters} can be exposed for both
\href{https://weather.aims.gov.au/#/overview}{Weather} and
\href{https://weather.aims.gov.au/#/overview}{Sea Water Temperature Loggers}
using function \code{\link{expose_attributes}}.

Note that at present the user can inspect the range of dates for
the temperature loggers data only (see usage of argument \code{summary} in
\code{\link{aims_data}}). Details about available dates for each dataset
and time series can be accessed via Metadata on
\href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}.
We raise this caveat here because these time boundaries are very important;
data are collected at very small time intervals, so just a few days of
time interval can yield massive datasets. The query will return and error
if it reaches the system's memory capacity.
}
\seealso{
\code{\link{filter_values}}, \code{\link{page_data}},
\code{\link{aims_data}}
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
