% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{next_page_data_get}
\alias{next_page_data_get}
\title{Get some data and return a data frame}
\usage{
next_page_data_get(url, api_key = NULL)
}
\arguments{
\item{url}{Description of parameter}

\item{api_key}{Description of parameter}
}
\value{
A \code{\link[jsonlite]{fromJSON}} simplified \code{data.frame}.
}
\description{
General description here
}
\details{
Details about this function.
}
\seealso{
\code{\link{aims_data_get}}
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
