% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_data.R
\name{page_data}
\alias{page_data}
\title{Request data via the AIMS Data Platform API}
\usage{
page_data(
  doi,
  filters = NULL,
  api_key = NULL,
  summary = NA,
  aims_version = NA,
  verbose = FALSE
)
}
\arguments{
\item{doi}{A \href{https://www.doi.org/}{Digital Object Identifier}
for a chosen
\href{https://open-aims.github.io/data-platform}{AIMS data series}}

\item{filters}{A \code{\link[base]{list}} containing a set of
filters for the data query (see Details).}

\item{api_key}{An AIMS Data Platform
\href{https://open-aims.github.io/data-platform/key-request}{API Key}}

\item{summary}{Should summary tables (\code{"summary-by-series"} or
\code{"summary-by-deployment"}) or daily aggregated data ("daily") be
returned instead of full data (see Details)?}

\item{aims_version}{A \code{\link[base]{character}} string
defining the version of database. Must be "/v1.0" or "-v2.0".
If none is provided, then "-v2.0" (the most recent) is used.}

\item{verbose}{Should links be printed to screen? Used for debugging only}
}
\value{
\code{aims_data} returns a \code{\link[base]{data.frame}} of class
\code{\link{aimsdf}}.

If \code{summary \%in\% c("summary-by-series", "summary-by-deployment")},
the output shows the summary information for the target dataset (i.e.
weather or temperature loggers)
(NB: currently, \code{summary} only works for the temperature logger
database). If \code{summary} is \emph{not} passed as an additional argument, then
the output contains \strong{raw} monitoring data. If \code{summary = "daily"},
then the output contains \strong{mean daily aggregated} monitoring data.
The output also contains five attributes (empty strings if
\code{summary} is passed as an additional argument):
\itemize{
\item{\code{metadata}}{a \href{https://www.doi.org/}{DOI} link
containing the metadata record for the data series.}
\item{\code{citation}}{the citation information for the particular
dataset.}
\item{\code{parameters}}{The measured parameters comprised in the
output.}
\item{\code{type}}{The type of dataset. Either "monitoring" if
\code{summary} is not specified, "monitoring (daily aggregation)" if
\code{summary = "daily"}, or a "summary-by-" otherwise.}
\item{\code{target}}{The input target.}
}
}
\description{
A function that communicates with the
the \href{https://open-aims.github.io/data-platform}{AIMS Data Platform}
via the AIMS Data Platform API
}
\details{
The AIMS Data Platform R Client provides easy access to
data sets for R applications to the
\href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}.
The AIMS Data Platform requires an API Key for requests, which can
be obtained at this
\href{https://open-aims.github.io/data-platform/key-request}{link}.
It is preferred that API Keys are not stored in code. We recommend
storing the environment variable \code{AIMS_DATAPLATFORM_API_KEY}
permanently under the user's \code{.Renviron} file in order to load
the API Key automatically.

There are two types of data currently available through the
\href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}:
\href{https://weather.aims.gov.au/#/overview}{Weather} and
\href{https://tinyurl.com/h93mcojk}{Sea Water Temperature Loggers}.
They are searched internally via unique DOI identifiers.
Only one data type at a time can be passed to the argument \code{target}.

A list of arguments for \code{filters} can be exposed for both
\href{https://weather.aims.gov.au/#/overview}{Weather} and
\href{https://weather.aims.gov.au/#/overview}{Sea Water Temperature Loggers}
using function \code{\link{aims_expose_attributes}}.

Note that at present the user can inspect the range of dates for
the temperature loggers data only (see usage of argument \code{summary} in
the examples below). For that, the argument \code{summary} must be either
the string \code{"summary-by-series"} or \code{"summary-by-deployment"}.
In those cases, time filters will be ignored.

Details about available dates for each dataset and time series can be
accessed via Metadata on
\href{https://open-aims.github.io/data-platform}{AIMS Data Platform API}.
We raise this caveat here because these time boundaries are very important;
data are collected at very small time intervals, a window of just a few days
can yield very large datasets. The query will return and error
if it reaches the system's memory capacity.

For that same reason, from version 1.1.0 onwards, we are offering the
possibility of downloading a mean daily aggregated version. For that, the
user must set \code{summary = "daily"}. In this particular case, query filter
will be taken into account.
}
\seealso{
\code{\link{aims_expose_attributes}},
\code{\link{aims_filter_values}}, \code{\link{aims_data}}
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
