% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{process_request}
\alias{process_request}
\title{Format \code{\link{json_results}} output}
\usage{
process_request(dt_req, next_page = FALSE, ...)
}
\arguments{
\item{dt_req}{An URL \code{\link[httr]{GET}} output}

\item{next_page}{Logical. Is this a multi-url request?}

\item{...}{Additional arguments to be passed to internal function
\code{\link{update_format}}}
}
\value{
\code{aims_data} returns a \code{\link[base]{data.frame}} of class
\code{\link{aimsdf}}.

If \code{summary} is passed as an additional argument,
the output shows the summary information for the target dataset (i.e.
weather or temperature loggers)
(NB: currently, \code{summary} only works for the temperature logger
database). If \code{summary} is \emph{not} passed as an additional argument, then
the output contains monitoring data.
The output also contains five attributes (empty strings if
\code{summary} is passed as an additional argument):
\itemize{
\item{\code{metadata}}{a \href{https://www.doi.org/}{DOI} link
containing the metadata record for the data series.}
\item{\code{citation}}{the citation information for the particular
dataset.}
\item{\code{parameters}}{The measured parameters comprised in the
output.}
\item{\code{type}}{The type of dataset. Either "monitoring" if
\code{summary} is not specified, or a "summary-by-" otherwise.}
\item{\code{target}}{The input target.}
}
}
\description{
Wrapper function
}
\details{
This function checks for errors in \code{dt_req}
data request and processes result via
\code{\link{json_results}}.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
\keyword{internal}
