% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataaimsr.R
\name{request_process}
\alias{request_process}
\title{Title here}
\usage{
request_process(dt_req, next_page = FALSE)
}
\arguments{
\item{dt_req}{Description of parameter}

\item{next_page}{Description of parameter}
}
\value{
A \code{\link[jsonlite]{fromJSON}} simplified \code{data.frame}.
}
\description{
General description here
}
\details{
Details about this function.
}
\author{
AIMS Datacentre \email{adc@aims.gov.au}
}
